slim down the Users class
